requirements.txt
google-adk[fastapi,ui]
python-dotenv
google-cloud-aiplatform
google-genai
uvicorn
